# django_project/python-flask-server/swagger_server/models/assignproject_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class AssignprojectRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, project_id: int=None, user_list: List[int]=None):  # noqa: E501
        """AssignprojectRequest - a model defined in Swagger

        :param project_id: The project_id of this AssignprojectRequest.  # noqa: E501
        :type project_id: int
        :param user_list: The user_list of this AssignprojectRequest.  # noqa: E501
        :type user_list: List[int]
        """
        self.swagger_types = {
            'project_id': int,
            'user_list': List[int]
        }

        self.attribute_map = {
            'project_id': 'project_id',
            'user_list': 'user_list'
        }

        self._project_id = project_id
        self._user_list = user_list

    @classmethod
    def from_dict(cls, dikt) -> 'AssignprojectRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AssignprojectRequest of this AssignprojectRequest.  # noqa: E501
        :rtype: AssignprojectRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def project_id(self) -> int:
        """Gets the project_id of this AssignprojectRequest.


        :return: The project_id of this AssignprojectRequest.
        :rtype: int
        """
        return self._project_id

    @project_id.setter
    def project_id(self, project_id: int):
        """Sets the project_id of this AssignprojectRequest.


        :param project_id: The project_id of this AssignprojectRequest.
        :type project_id: int
        """

        self._project_id = project_id

    @property
    def user_list(self) -> List[int]:
        """Gets the user_list of this AssignprojectRequest.


        :return: The user_list of this AssignprojectRequest.
        :rtype: List[int]
        """
        return self._user_list

    @user_list.setter
    def user_list(self, user_list: List[int]):
        """Sets the user_list of this AssignprojectRequest.


        :param user_list: The user_list of this AssignprojectRequest.
        :type user_list: List[int]
        """

        self._user_list = user_list