# django_project/python-flask-server/swagger_server/models/createproject_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CreateprojectRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, project_name: str=None):  # noqa: E501
        """CreateprojectRequest - a model defined in Swagger

        :param project_name: The project_name of this CreateprojectRequest.  # noqa: E501
        :type project_name: str
        """
        self.swagger_types = {
            'project_name': str
        }

        self.attribute_map = {
            'project_name': 'project_name'
        }

        self._project_name = project_name

    @classmethod
    def from_dict(cls, dikt) -> 'CreateprojectRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateprojectRequest of this CreateprojectRequest.  # noqa: E501
        :rtype: CreateprojectRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def project_name(self) -> str:
        """Gets the project_name of this CreateprojectRequest.


        :return: The project_name of this CreateprojectRequest.
        :rtype: str
        """
        return self._project_name

    @project_name.setter
    def project_name(self, project_name: str):
        """Sets the project_name of this CreateprojectRequest.


        :param project_name: The project_name of this CreateprojectRequest.
        :type project_name: str
        """

        self._project_name = project_name