# django_project/python-flask-server/swagger_server/models/createuser_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CreateuserRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, user_name: str=None):  # noqa: E501
        """CreateuserRequest - a model defined in Swagger

        :param user_name: The user_name of this CreateuserRequest.  # noqa: E501
        :type user_name: str
        """
        self.swagger_types = {
            'user_name': str
        }

        self.attribute_map = {
            'user_name': 'user_name'
        }

        self._user_name = user_name

    @classmethod
    def from_dict(cls, dikt) -> 'CreateuserRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateuserRequest of this CreateuserRequest.  # noqa: E501
        :rtype: CreateuserRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_name(self) -> str:
        """Gets the user_name of this CreateuserRequest.


        :return: The user_name of this CreateuserRequest.
        :rtype: str
        """
        return self._user_name

    @user_name.setter
    def user_name(self, user_name: str):
        """Sets the user_name of this CreateuserRequest.


        :param user_name: The user_name of this CreateuserRequest.
        :type user_name: str
        """

        self._user_name = user_name