# django_project/python-flask-server/swagger_server/models/project_assignmentor_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ProjectAssignmentorRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, mentor_id: int=None, project_id: int=None):  # noqa: E501
        """ProjectAssignmentorRequest - a model defined in Swagger

        :param mentor_id: The mentor_id of this ProjectAssignmentorRequest.  # noqa: E501
        :type mentor_id: int
        :param project_id: The project_id of this ProjectAssignmentorRequest.  # noqa: E501
        :type project_id: int
        """
        self.swagger_types = {
            'mentor_id': int,
            'project_id': int
        }

        self.attribute_map = {
            'mentor_id': 'mentor_id',
            'project_id': 'project_id'
        }

        self._mentor_id = mentor_id
        self._project_id = project_id

    @classmethod
    def from_dict(cls, dikt) -> 'ProjectAssignmentorRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProjectAssignmentorRequest of this ProjectAssignmentorRequest.  # noqa: E501
        :rtype: ProjectAssignmentorRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def mentor_id(self) -> int:
        """Gets the mentor_id of this ProjectAssignmentorRequest.


        :return: The mentor_id of this ProjectAssignmentorRequest.
        :rtype: int
        """
        return self._mentor_id

    @mentor_id.setter
    def mentor_id(self, mentor_id: int):
        """Sets the mentor_id of this ProjectAssignmentorRequest.


        :param mentor_id: The mentor_id of this ProjectAssignmentorRequest.
        :type mentor_id: int
        """

        self._mentor_id = mentor_id

    @property
    def project_id(self) -> int:
        """Gets the project_id of this ProjectAssignmentorRequest.


        :return: The project_id of this ProjectAssignmentorRequest.
        :rtype: int
        """
        return self._project_id

    @project_id.setter
    def project_id(self, project_id: int):
        """Sets the project_id of this ProjectAssignmentorRequest.


        :param project_id: The project_id of this ProjectAssignmentorRequest.
        :type project_id: int
        """

        self._project_id = project_id